requirements.txt
pyexiftool
python-dateutil